fix : 경로

+++ src/main/java/store/buzzbook/core/dto/product/response/BookResponse.java
@@ -1,5 +1,7 @@
 package store.buzzbook.core.dto.product.response;
 
+import java.time.ZonedDateTime;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -7,28 +9,26 @@ import lombok.NoArgsConstructor;
 import store.buzzbook.core.entity.product.Book;
 import store.buzzbook.core.entity.product.Publisher;
 
-import java.time.ZonedDateTime;
-
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
 @Builder
 public class BookResponse {
-    private int id;
-    private String title;
-    private String description;
-    private String isbn;
-    private Publisher publisher;
-    private ZonedDateTime publishDate;
+	private long id;
+	private String title;
+	private String description;
+	private String isbn;
+	private Publisher publisher;
+	private ZonedDateTime publishDate;
 
-    public static BookResponse convertToBookResponse(Book book) {
-        return BookResponse.builder()
-                .id(book.getId())
-                .title(book.getTitle())
-                .description(book.getDescription())
-                .isbn(book.getIsbn())
-                .publisher(book.getPublisher())
-                .publishDate(book.getPublishDate())
-                .build();
-    }
+	public static BookResponse convertToBookResponse(Book book) {
+		return BookResponse.builder()
+			.id(book.getId())
+			.title(book.getTitle())
+			.description(book.getDescription())
+			.isbn(book.getIsbn())
+			.publisher(book.getPublisher())
+			.publishDate(book.getPublishDate())
+			.build();
+	}
 }
